(#958) Don't use compound names for decorators
Fixed packages io and iterable.
