--- src/test/java/org/cactoos/iterable/StickyIterableTest.java
@@ -1,56 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2017-2018 Yegor Bugayenko
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.cactoos.iterable;
-
-import java.util.Collections;
-import java.util.concurrent.atomic.AtomicInteger;
-import org.cactoos.list.ListOf;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.Test;
-
-/**
- * Test case for {@link StickyIterable}.
- *
- * @since 0.8
- * @checkstyle JavadocMethodCheck (500 lines)
- */
-public final class StickyIterableTest {
-
-    @Test
-    public void ignoresChangesInIterable() throws Exception {
-        final AtomicInteger size = new AtomicInteger(2);
-        final Iterable<Integer> list = new StickyIterable<>(
-            new ListOf<>(
-                () -> Collections.nCopies(size.incrementAndGet(), 0).iterator()
-            )
-        );
-        MatcherAssert.assertThat(
-            "Can't ignore the changes in the underlying iterable",
-            new LengthOf(list).intValue(),
-            Matchers.equalTo(new LengthOf(list).intValue())
-        );
-    }
-
-}
